% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFile.OSCURS.R
\name{parseFile.OSCURS}
\alias{parseFile.OSCURS}
\title{Parse an OSCURS output trajectory file to a dataframe}
\usage{
parseFile.OSCURS(fn, verbose = FALSE)
}
\arguments{
\item{fn}{\itemize{
\item name of file to parse
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a list with elements "dfr" and "track". The latter is a spatial tibble using a sf line class to represent the track.
}
\description{
Function to parse an OSCURS output trajectory file to a dataframe.
}
\details{
Parses output text file from OSCURS run.  The returned object is a list with elements "dfr" and "track". 'dfr' is a dataframe with
each row representing a time and position along the track. "track is a spatial tibble using an
sf line geometry class to represent the track.

Requires packages \code{stringr}, \code{tibble} and \code{tmap_tools}.
}
