% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPTI.R
\name{plotPTI}
\alias{plotPTI}
\title{Plot particle tracks from OSCURS model runs on a map for the PAPA Trajectory Index}
\usage{
plotPTI(tracks, text_size = 1, strCRS = tmaptools::get_proj4(3338,
  output = "character"), basemap = wtsGIS::createBaseTMap(layer.land =
  wtsGIS::getPackagedLayer("Alaska"), layer.bathym =
  wtsGIS::getPackagedLayer("ShelfBathymetry"), strCRS.finl = strCRS),
  bounding_box = tmaptools::bb(xlim = c(-165, -135), ylim = c(48, 59),
  current.projection = tmaptools::get_proj4("longlat", output =
  "character"), projection = strCRS), gridLines = list(x = seq(from =
  -165, to = -135, by = 5), y = seq(from = 48, to = 59, by = 2), projection
  = strWGS84), style = "white", color = "blue", alpha = 1,
  palette = NULL, showMap = TRUE, verbose = FALSE)
}
\arguments{
\item{tracks}{- a tibble with OSCURS tracks (e.g., the \code{tracks} list element from a call to \code{convertOSCURStoTbl})}

\item{text_size}{- size for "year" label on last end point (others will be 0.8*text_size)}

\item{strCRS}{- character representation of coordinate reference system for final map (default is Alaska Albers [EPSG=3338])}

\item{basemap}{- a base map for plotting the tracks (default is the EBS using CRS defined by strCRS)}

\item{bounding_box}{- bounding box for map (or NULL to calculate from track limits)}

\item{gridLines}{- list w/ info to plot grid lines on map (or NULL for no grid lines)}

\item{color}{- color for track lines (or column name)}

\item{alpha}{- transparency for track lines}

\item{showMap}{- flag to show the map}

\item{verbose}{- flag to print diagnostic info}
}
\value{
- a list (see details)
}
\description{
Function to plot particle tracks from OSCURS model runs on a map for the PAPA Trajectory Index.
}
\details{
The returned list has the following elements:
 * map       - a tmap object
 * tracks    - a sf tibble with tracks that span the IDL to "roll your own" map
 * startLocs - a sf tibble with starting track locations as points to "roll your own" map
 * endLocs - a sf tibble with ending track locations as points to "roll your own" map
 * basemap   - a basemap to "roll your own" map

The map object is in the coordinate reference system given by strCRS.

Requires packages \code{dplyr}, \code{magrittr}, \code{sf}, \code{tmap}, \code{wtsGIS}.
}
