% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOSCURStoTbl.R
\name{convertOSCURStoTbl}
\alias{convertOSCURStoTbl}
\title{Collates sets of OSCURS model output to a list with a data.frame and a spatial tibble of type sf}
\usage{
convertOSCURStoTbl(fnBase = "./OSCURS_", stYrs = 2017,
  stMDs = list(APR = 15, MAY = c(1, 15), JUN = 1), stLLs = NULL,
  verbose = FALSE)
}
\arguments{
\item{fnBase}{- base for OSCURS output filenames to process (default = "./OSCURS_")}

\item{stYrs}{- start years for OSCURS files (vector)}

\item{stMDs}{- start months/days for OSCURS files (list)}

\item{stLLs}{- file name for csv with table of starting locations for particle releases}

\item{verbose}{- flag to print diagnostic output}
}
\value{
list with elements "dfr" (a data.frame) and "tracks" (a spatial tibble using sf classes)
}
\description{
A function to collate sets of OSCURS model output created using \code{runOSCURS}
to a list with a data.frame and a spatial tibble of type sf.
}
\details{
Requires packages \code{magrittr} and \code{sf}.
Note that the coordinate reference system assigned to
tbl.track is WGS84, with longitude running -180 to 180 (EPSG=4326).
}
