% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doOSCURS.R
\name{doOSCURS}
\alias{doOSCURS}
\title{Run OSCURS and plot results}
\usage{
doOSCURS(fnBase = "OSCURS_", path = ".", nDays = 90, stYrs = 2017,
  stMDs = list(APR = 15), fnStartLocs = "OSCURS_StartLocations.csv",
  wcsc = 1, wad = 0, gsf = 1,
  link = "https://oceanview.pfeg.noaa.gov/oscurs/runOscurs9.php?",
  randNum = round(runif(1, 1, 1e+05)),
  strCRS = tmaptools::get_proj4(3338, output = "character"),
  basemap = wtsGIS::createBaseTMap(layer.land =
  wtsGIS::getPackagedLayer("Alaska"), layer.bathym =
  wtsGIS::getPackagedLayer("ShelfBathymetry"), strCRS.finl = strCRS),
  alpha = 1, showMap = TRUE, verbose = FALSE)
}
\arguments{
\item{fnBase}{- base for output filenames (default = "OSCURS_")}

\item{path}{- path to folder in which to store output files}

\item{nDays}{- number of days to run model}

\item{stYrs}{- vector of years for particle releases}

\item{stMDs}{- list of release dates, by month}

\item{fnStartLocs}{- filename for csv file with initial particle lat/lon locations}

\item{wcsc}{- wind/current speed coefficient (default=1.0; see https://oceanview.pfeg.noaa.gov/oscurs/)}

\item{wad}{- wind angle deviation (default=0.0; see https://oceanview.pfeg.noaa.gov/oscurs/)}

\item{gsf}{- geostrophic speed factor (default=1.0; see https://oceanview.pfeg.noaa.gov/oscurs/)}

\item{link}{- url to run OSCURS (default = "https://oceanview.pfeg.noaa.gov/oscurs/runOscurs9.php?")}

\item{randNum}{- random number to add to server-side file name to ensure uniqueness}

\item{strCRS}{- character representation of coordinate reference system for final map (default is Alaska Albers [EPSG=3338])}

\item{basemap}{- a base map for plotting the tracks (default is the EBS using CRS defined by strCRS)}

\item{alpha}{- transparency for track lines}

\item{showMap}{- flag to show the map}

\item{verbose}{- flag to print diagnostic info}
}
\value{
same list as \code{plotOSCURS} (see @details).
}
\description{
Function to run OSCURS and plot results.
}
\details{
Requires packages \code{readr}, \code{tmaptools}, \code{wtsGIS}.
The returned list has the following elements:
 * map       - a tmap object
 * tracks    - a sf tibble with tracks that span the IDL to "roll your own" map
 * startLocs - a sf tibble with starting track locations as points to "roll your own" map
 * basemap   - a basemap to "roll your own" map
}
