% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFile.OSCURS.R
\name{parseFile.OSCURS}
\alias{parseFile.OSCURS}
\title{Parse an OSCURS output trajectory file to a dataframe}
\usage{
parseFile.OSCURS(fn, verbose = FALSE)
}
\arguments{
\item{fn}{- name of file to parse}

\item{verbose}{- flag to print diagnostic info}
}
\value{
list with elements "dfr" and "track". The latter is a spatial tibble using sf classes.
}
\description{
Function to parse an OSCURS output trajectory file to a dataframe.
}
\details{
Parses output text file from OSCURS run. Requires packages \code{stringr}, \code{tibble} and
\code{tmap_tools}.
}
