% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOSCURS.R
\name{plotOSCURS}
\alias{plotOSCURS}
\title{Plot particle tracks from OSCURS model runs on a map}
\usage{
plotOSCURS(tracks, stLLs, strCRS = tmaptools::get_proj4(3338, output =
  "character"), basemap = wtsGIS::createBaseTMap(layer.land =
  wtsGIS::getPackagedLayer("Alaska"), layer.bathym =
  wtsGIS::getPackagedLayer("ShelfBathymetry"), strCRS.finl = strCRS),
  alpha = 1, showMap = TRUE, verbose = FALSE)
}
\arguments{
\item{tracks}{- a tibble with OSCURS tracks (e.g., the \code{tracks} list element from a call to \code{convertOSCURStoTbl})}

\item{stLLs}{- a tibble with starting particle locations (e.g., from reading the initial particle locations csv file using \code{readr::read_csv})}

\item{strCRS}{- character representation of coordinate reference system for final map (default is Alaska Albers [EPSG=3338])}

\item{basemap}{- a base map for plotting the tracks (default is the EBS using CRS defined by strCRS)}

\item{alpha}{- transparency for track lines}

\item{showMap}{- flag to show the map}

\item{verbose}{- flag to print diagnostic info}
}
\value{
- a tmap object
}
\description{

}
\details{
The map object is in the Alaska Albers projection.
Requires packages \code{dply}, \code{magrittr}, \code{sf}, \code{tmap}, \code{wtsGIS}.
}
