% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOSCURS.R
\name{runOSCURS}
\alias{runOSCURS}
\title{Make a series of OSCURS runs}
\usage{
runOSCURS(fnBase = "OSCURS_", path = ".", nDays = 90, stYrs = 2017,
  stMDs = list(APR = 15, MAY = c(1, 15), JUN = 1), stLLs = NULL,
  link = "https://oceanview.pfeg.noaa.gov/oscurs/runOscurs9.php?",
  wcsc = 1, wad = 0, gsf = 1, randNum = round(runif(1, 1, 1e+05)),
  test = TRUE, verbose = FALSE)
}
\arguments{
\item{fnBase}{- base for filenames (default = "OSCURS_")}

\item{path}{- path to use with output filenames (default = ".")}

\item{nDays}{- number of days to run model}

\item{stYrs}{- vector of years for particle releases}

\item{stMDs}{- list of release dates, by month}

\item{stLLs}{- a dataframe or tibble with initial particle lat/lon locations (see details below)}

\item{link}{- url to run OSCURS (default = "https://oceanview.pfeg.noaa.gov/oscurs/runOscurs9.php?")}

\item{wcsc}{- wind/current speed coefficient (default=1.0; see https://oceanview.pfeg.noaa.gov/oscurs/)}

\item{wad}{- wind angle deviation (default=0.0; see https://oceanview.pfeg.noaa.gov/oscurs/)}

\item{gsf}{- geostrophic speed factor (default=1.0; see https://oceanview.pfeg.noaa.gov/oscurs/)}

\item{randNum}{- random number to add to server-side file name to ensure uniqueness}

\item{test}{- flag to print diagnostic output but NOT run OSCURS (useful to see what commands will be sent)}

\item{verbose}{- flag to print diagnostic output}
}
\description{
Function to make a series of OSCURS runs and download the results to a set of files.
}
\details{
Runs the OSCURS model at the link given. An output file is created for each particle release of the
form fnBase_year_month_day_latdegree_latminute_londegree_lonminute.csv, where lat and lon is the initial
particle location.

The input parameter \code{stLLs} should be a dataframe or tibble. The following columns are required:
 * LATITUDE   - latitude in decimal degrees (-90 to 90)
 * LONGITUDE  - longitude in decimal degrees (-180 to 180)
}
