% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOSCURS.R
\name{runOSCURS}
\alias{runOSCURS}
\title{Make a set of OSCURS runs}
\usage{
runOSCURS(fnBase = "./OSCURS_", nDays = 90, stYrs = 2017,
  stMDs = list(APR = 15, MAY = c(1, 15), JUN = 1), stLLs = NULL,
  link = "https://oceanview.pfeg.noaa.gov/oscurs/runOscurs9.php?",
  test = TRUE, verbose = FALSE)
}
\arguments{
\item{fnBase}{- base for output filenames (default = "./OSCURS_")}

\item{nDays}{- number of days to run model}

\item{stYrs}{- vector of years for particle releases}

\item{stMDs}{- list of release dates, by month}

\item{stLLs}{- a tibble with initial particle lat/lon locations}

\item{link}{- url to run OSCURS (default = "https://oceanview.pfeg.noaa.gov/oscurs/runOscurs9.php?")}

\item{test}{- flag to print diagnostic output but NOT run OSCURS (useful to see what commands will be sent)}

\item{verbose}{- flag to print diagnostic output}
}
\description{
Function to make a series of OSCURS runs
}
\details{
Runs the OSCURS model at the link given. An output file is created for each particle release of the
form fnBase_year_month_day_latdegree_latminute_londegree_lonminute.csv, where lat and lon is the initial
particle location.
}
